Switch from Maven to Gradle (to have proper microbench building w/o making jamm a multi-module-project just to have microbenchmarks and to not keep microbenchmarks in the main source tree)
The main reason to switch from Maven to Gradle is the microbenchmark, i.e. with
Maven the published pom.xml still contains the jmh dependencies and the production
jamm.jar file the Microbench.class.

It would have been possible to migrate jamm to a multi-module project to get around
the dependency issues w/ Maven, but since jamm is such a small project, it felt
easier to migrate to Gradle.