Add support for the Module System introduced in Java 9
This patch avoid IllegalReflectiveAccessExceptions by relying on Unsafe to
access objects that should not be accessible from the unnamed module.
Unfortunately, even if the code ensure that it will not do any illegal
reflective access the design of the JVM is such that we cannot avoid some
warnings about future illegal access.